changes as per the comments
