+++ example.env
@@ -1,3 +1,2 @@
 WALLET_PRIVATE_KEY=0x...
 CHAIN_ID=
-DATA_API_KEY=
