--- src/sdk/primeDataUtils.ts
@@ -1,183 +1,0 @@
-import "reflect-metadata";
-import { AccountBalances, AdvanceRoutesLiFi, NftList, PaginatedTokens, PrimeDataModule, RateData, StepTransactions, TokenList, TokenListToken, Transaction } from "./data";
-import { FetchExchangeRatesDto, GetAccountBalancesDto, GetAdvanceRoutesLiFiDto, GetExchangeSupportedAssetsDto, GetNftListDto, GetStepTransactionsLiFiDto, GetTokenListDto, GetTokenListsDto, GetTransactionDto, validateDto } from "./dto";
-import { BigNumber } from "ethers";
-
-export class PrimeDataUtils {
-    private primeDataModule: PrimeDataModule;
-    constructor(apiKey: string) {
-        this.primeDataModule = new PrimeDataModule(apiKey)
-    }
-
-    /**
-    * gets account balances
-    * @param dto
-    * @return Promise<AccountBalances>
-    */
-    async getAccountBalances(dto: GetAccountBalancesDto): Promise<AccountBalances> {
-        const { account, tokens, chainId, provider } = await validateDto(dto, GetAccountBalancesDto, {
-            addressKeys: ['account', 'tokens'],
-        });
-
-        return this.primeDataModule.getAccountBalances(
-            account,
-            tokens,
-            chainId,
-            provider,
-        );
-    }
-
-    /**
-    * gets transaction
-    * @param dto
-    * @return Promise<Transaction>
-    */
-    async getTransaction(dto: GetTransactionDto): Promise<Transaction> {
-        const { hash, chainId } = await validateDto(dto, GetTransactionDto);
-
-        return this.primeDataModule.getTransaction(hash, chainId);
-    }
-
-    /**
-    * gets NFT list belonging to account
-    * @param dto
-    * @return Promise<NftList>
-    */
-    async getNftList(dto: GetNftListDto): Promise<NftList> {
-        const { account, chainId } = await validateDto(dto, GetNftListDto, {
-            addressKeys: ['account'],
-        });
-
-        return this.primeDataModule.getNftList(
-            account,
-            chainId,
-        );
-    }
-
-    /**
-    * gets advance routes from LIFI
-    * @param dto
-    * @return Promise<AdvanceRoutesLiFi>
-    */
-    async getAdvanceRoutesLiFi(dto: GetAdvanceRoutesLiFiDto): Promise<AdvanceRoutesLiFi> {
-        const {
-            fromChainId,
-            toChainId,
-            fromTokenAddress,
-            toTokenAddress,
-            fromAmount,
-            allowSwitchChain,
-            showZeroUsd,
-            fromAddress,
-        } = await validateDto(dto, GetAdvanceRoutesLiFiDto, {
-            addressKeys: ['fromTokenAddress', 'toTokenAddress', 'fromAddress'],
-        });
-
-        let { toAddress } = dto;
-
-        if (!toAddress) toAddress = fromAddress;
-
-        const data = await this.primeDataModule.getAdvanceRoutesLiFi(
-            fromTokenAddress,
-            toTokenAddress,
-            fromChainId,
-            toChainId,
-            BigNumber.from(fromAmount),
-            toAddress,
-            allowSwitchChain,
-            fromAddress,
-            showZeroUsd,
-        );
-
-        return data;
-    }
-
-    /**
-    * gets step transactions from LIFI
-    * @param dto
-    * @return Promise<StepTransactions>
-    */
-    async getStepTransaction(dto: GetStepTransactionsLiFiDto): Promise<StepTransactions> {
-        const { route, account } = await validateDto(dto, GetStepTransactionsLiFiDto, {
-            addressKeys: ['account']
-        })
-
-        return this.primeDataModule.getStepTransaction(route, account);
-    }
-
-    /**
-    * gets exchange supported tokens
-    * @param dto
-    * @return Promise<PaginatedTokens>
-    */
-    async getExchangeSupportedAssets(dto: GetExchangeSupportedAssetsDto): Promise<PaginatedTokens> {
-        const { page, limit, chainId, account } = await validateDto(dto, GetExchangeSupportedAssetsDto, {
-            addressKeys: ['account']
-        });
-
-        return this.primeDataModule.getExchangeSupportedAssets(page, limit, chainId, account);
-    }
-
-    /**
-    * gets token lists
-    * @param dto
-    * @return Promise<TokenList[]>
-    */
-    async getTokenLists(dto: GetTokenListsDto): Promise<TokenList[]> {
-        const { chainId } = await validateDto(dto, GetTokenListsDto);
-
-        return this.primeDataModule.getTokenLists(chainId);
-    }
-
-    /**
-    * gets token list tokens
-    * @param dto
-    * @return Promise<TokenListToken[]>
-    */
-    async getTokenListTokens(dto: GetTokenListDto): Promise<TokenListToken[]> {
-        const { chainId, name } = await validateDto(dto, GetTokenListDto);
-
-        return this.primeDataModule.getTokenListTokens(chainId, name);
-    }
-
-    /**
-    * fetch exchange rates of tokens
-    * @param dto
-    * @return Promise<RateData>
-    */
-    async fetchExchangeRates(dto: FetchExchangeRatesDto): Promise<RateData> {
-        const { tokens, chainId } = dto;
-        let data: RateData;
-        const promises = [];
-
-        // Create a batch of 50
-        const batches = [...Array(Math.ceil(tokens.length / 50))].map(() => tokens.splice(0, 50));
-        batches.forEach((batch) => {
-            promises.push(this.primeDataModule.fetchExchangeRates(batch, chainId));
-        });
-
-        // Fetch succeded results and merge
-        await (Promise as any)
-            .allSettled(promises)
-            .then((response) =>
-                response?.forEach((result) => {
-                    if (result?.status === 'fulfilled') {
-                        !data
-                            ? (data = result.value ? result.value : {})
-                            : (data.items = result?.value?.items ? [...data.items, ...result.value.items] : [...data.items]);
-                    }
-                }),
-            );
-
-        // Return Unique tokens
-        if (data && data.items && data.items.length) {
-            data.error = ''
-            data.errored = false
-            data.items = [...new Map(data.items.map(item => [item['address'], item])).values()];
-        } else {
-            data.items = [];
-        }
-
-        return data;
-    }
-}
